% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4rvendor-package.R
\name{cpp4rvendor-package}
\alias{cpp4rvendor-package}
\title{To Test Vendored Builds}
\description{
Testing the builds with vendored cpp4r headers. The idea is to
 ensure that the functions correctly link to the vendored cpp4r code and not
 to any installed version of cpp4r.
}

