% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4rvendor-package.R
\docType{package}
\name{cpp4rvendor-package}
\alias{cpp4rvendor}
\alias{cpp4rvendor-package}
\title{cpp4rvendor: To test vendored buils}
\description{
Testing the builds with vendored cpp4r headers. The idea is to ensure that the functions correctly link to the vendored cpp4r code and not to any installed version of cpp4r.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://WEBSITE.COM}
  \item Report bugs at \url{https://github.com/USERNAME/PKGNAME/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
